imx8mp gpc module addto kernel and project restructure for easy development on MSVC
GPC module is responsible for power up/down peripherals present on imx8mp board,
gpc module is added to kernel.
Project is being restructured for MSVC for easy development and to gain github run check success.

Manas Kamal Choudhury

+++ KernelAA64/Mm/vmmngr.c
@@ -520,7 +520,7 @@ void AuUpdatePageFlags(uint64_t virt_addr, uint64_t flags) {
 
 	//AuTextOut("Your physical page is -> %x %x\n", page, V2P(page));
 	if (page) {
-		pt[pt_index(virt_addr)] = (V2P(page) & ~0xFFFULL) | flags;
+		pt[pt_index(virt_addr)] = (V2P((uint64_t)page) & ~0xFFFULL) | flags;
 		data_cache_flush(&pt[pt_index(virt_addr)]);
 	}
 }
